CIFTItoFC

Extracting FC matrices from CIFTI files

CIFTItoFC(
  path = "./",
  wb_path = "/home/junhong.yu/workbench/bin_rh_linux64",
  dtseries = "_space-fsLR_den-91k_desc-denoised_bold.dtseries.nii",
  timeseries = F,
  concat_subj = TRUE,
  atlas = 200,
  filename = "FC.rds"
)

Arguments

path

The filepath to directory containing the subject folders. Set to ./. by default.

wb_path

The filepath to the workbench folder that you have previously downloaded and unzipped. Set to /home/junhong.yu/workbench/bin_rh_linux64 by default

dtseries

The filename extension of the fMRI volumes. Set to _space-fsLR_den-91k_desc-denoised_bold.dtseries.nii by default

timeseries

If set to TRUE, FC matrices will not be computed, instead the parcellated timeseries will be return in a list format, where each subject's parcellated timeseries data is contained within a list element. Set to FALSE by default.

concat_subj

When set to TRUE (default), timeseries data from multiple runs/sessions of the same subject is first Z-scaled and then concatenated into a single larger timeseries data frame before computing the FC.

atlas

The version (number of parcels) of the Schaefer atlas; it has to be in multiples of 100, from 100 to 1000. set to 200 by default. The specified atlas template will be automatically downloaded from here if it does not exist in the current directory.

filename

Filename of the concatenated FC vector file. Set to FC.rds by default

Value

outputs N (number of subjects) x E (number of unique) matrices as a .rds file

Details

This function extracts FC matrices from CIFTI volumes postprocessed with XCP_d

Examples

if (FALSE) { # \dontrun{
CIFTItoFC(wb_path="bin_windows64/",dtseries="_ses-0._task-rest_acq-.._run-0._space-fsLR_den-91k_desc-denoised_bold.dtseries.nii",atlas = 200,filename="test.RDS")
} # }