extractFC

Extracting FC matrices from HCP-preprocessed fMRI volumes

extractFC(
  wb_path,
  base_dir = "fmriresults01/",
  atlas,
  FD = F,
  motion.thresh = 0.25,
  scrub = F,
  GSR = F,
  task,
  extension = "_Atlas_MSMAll_hp0_clean.dtseries.nii",
  movement.extension = "Movement_RelativeRMS.txt",
  output_dir = "FCmat",
  report_filename = "report.csv"
)

Arguments

wb_path

The filepath to the workbench folder that you have previously downloaded and unzipped

base_dir

The filepath to directory containing the subject folders. Set to fmriresults01/" by default.

atlas

The version (number of parcels) of the Schaefer atlas; it has to be in multiples of 100, from 100 to 1000. The specified atlas template will be automatically downloaded from here if it does not exist in the current directory.

FD

option to use FD instead of RMSD for measuring head motion. Set to FALSE by default.

motion.thresh

the threshold (Root Mean Squared displacement) used for scrubbing frames. Set to 0.25 by default.

scrub

If set to TRUE, remove frames with excessive head motion (relative RMS displacement> 0.25) and also remove segments in between two time-points if the segment has less than 5 frames. This is the ‘full scrubbing; method described in Power et. al (2014). This is not recommended for task-based fMRI volumes.

GSR

If set to TRUE, global signal and its square-term will be regressed out from the fMRI timeseries data. Note, that while the use of GSR enhances brain-behavior association, its use is controversial in the literature

task

The name of the task (case-sensitive), without any numbers or acquisition direction labels

extension

The filename extension of the fMRI volumes. Set to _Atlas_MSMAll_hp0_clean.dtseries.nii by default

movement.extension

The filename extension of the head motion files. Set to Movement_RelativeRMS.txt by default

output_dir

The output directory where the FC matrices will be saved

report_filename

The desired filename of the report containing columns of subj(subject ID),Mean_RMS(Mean Root Mean Squared displacement), and no_frames_removed(number of frames removed; only if scrub=TRUE)

Value

outputs M x M matrices in the output_dir and a report file (.csv format) in the working directory containing the head motion measurements

Details

This function extracts FC matrices from fMRI volumes preprocessed with the HCP pipeline with the option of carrying out global signal regression and scrubbing. The directory structure and file names must adhere to the NDA/HCP data structure and file naming conventions. Connectome Workbench has to be installed prior to running the function.

Examples

if (FALSE) {
extractFC(wb_path = "/home/junhong.yu/workbench/",task="CONFLICT",atlas=100,GSR=TRUE,scrub=TRUE,output_dir = "FCmat")
}